pyproject.toml
[project]
name = "proyecto-final-bigdata"
version = "0.1.0"
description = "Proyecto Final Big Data - Yelp Dataset - MAD UTP Verano 2026"
requires-python = ">=3.11"
dependencies = [
    "pyspark>=3.5.0",
    "pymongo>=4.6.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.2.0",
    "matplotlib>=3.8.0",
    "pypdf>=6.10.0",
    "seaborn>=0.13.0",
    "statsmodels>=0.14.0",
    "scikit-learn>=1.4.0",
]
